pyproject.toml
[project]
name = "introspection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "nnsight>=0.5.7",
    "python-dotenv>=1.2.1",
    "tokenizers>=0.13.0",
]
